Update Document Reference Integration with new demo doc images
Doc Ref Integration needed to be updated with new POA & POI
document images. Currently the same demo image is returned
and it is confusing/misleading as it is labelled for POA.

This change adds 4 new images (2 for both POA and POI) which
are labelled clearly and take into account if a check has
passed or failed.

Refactored _proof_of_address and _proof_of_identity into one
single _proof_document function. Logic added to check if a
document object is identity or address, and if its a pass or
fail result - this then adds a reference to document files
depending on what it is.

Updated the fields_from_result_type functions to include
all demo result types. Added to download_files function
to check a documents file reference and return the correct
image based on this.

Adding a new test file to test the _synthesize_demo_result
function in doccapture. Two tests added - one to test all pass
result, and the other to test a failed result with an uncertain
field - so we can also check the field checks are working
correctly.

Changes to the dockerfile to allow for local pytests to run.