Making changes requested via PR
This commit holds changes across a few files to improve and add
to changes which were made in the initial commit for this PR.
As requested by the reviewer.

Changed the test_download_files test in test_run_check.py so we
can test inital changes made in a previous commit. Now it will
test and check every document downloads the correct demo file.
I also changed the demo result which is passed so instead of
providing passing docs for all checks, one of them will fail.
This would make for a better test which is why I implimented it.

Removed lines which were added to the Dockerfile as these were
only intended to be used for testing locally.

Improved the logic when setting file ref in doccapture
_synthesize_demo_result so the code is cleaner.

Moved the static files to the doccapture directory as they are
relevant to this demo check only.

Improved tests written in test_demo_data.py to improve readability
by looking up the document category and field_check by their value
and not relying on list order.