Changes to the test file
This commit is for further changes to the tests for downloading
demo check files.

Changed the demo result passed to the demo check back to ALL_PASS.

Implimented the logic that looks up the documents by their category
rather than the list order to the test_download_files test. This
makes the test more readable.

Typo fix in _proof_of_document function.